backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
langgraph>=1.0.7
langchain-openai>=0.3.35
langchain-community>=0.3.0
langchain-core>=0.3.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
azure-communication-email>=1.0.0
sse-starlette>=2.0.0
pydantic>=2.0.0
